1. 记忆气泵区间若出错，只能重新烧录才能解决。风险较高，因此去除该功能。使用宽扫频代替 2. 优化头文件配置显示 3. 将采样数128替换成ADC_CH_SIZE，兼容20K~30K和100K的驱动程序 4. 修复小范围扫频结果在边界时，不能继续扫频的问题 5. 进水检测逻辑更改（待测试） 6. 软件版本升级到V1.8

+++ Core/Inc/usart.h
@@ -116,7 +116,7 @@ void usart_callback(void);
 
 /* 协议处理函数声明 */
 uint16_t crc16(uint8_t* buff, uint32_t len);
-uint8_t usart_send_frame(uint8_t cmd, uint8_t *data, uint16_t data_len);
+uint8_t usart_send_frame(uint8_t cmd, uint8_t *data, uint8_t data_len);
 void usart_handle_protocol_command(uint8_t *frame_data, uint16_t frame_len);
 void uart_cmd_process(void);
 void usart_transmit(uint8_t *buf, uint32_t len);
